FEAT: Implemented BotUI

+++ src/main/java/edu/bu/met622/database/MongoDB.java
@@ -118,8 +118,8 @@ public class MongoDB {
      * @param year    The year to search within
      * @return The number of times the keyword was found within the specified year
      */
-    public int query(String keyword, String year) {
-        int hits = 0;
+    public double query(String keyword, String year) {
+        double hits = 0;
 
         startTime = System.currentTimeMillis();                           // Start the runtime clock
 
@@ -149,8 +149,8 @@ public class MongoDB {
      * @param endYear    The last year within the search range
      * @return The number of times the keyword was found in the specified year
      */
-    public int query(String keyword, String startMonth, String startYear, String endMonth, String endYear) {
-        int hits = 0;
+    public double query(String keyword, String startMonth, String startYear, String endMonth, String endYear) {
+        double hits = 0;
         String startDate = startYear + "-" + startMonth + "-" + "1";
         String endDate = endYear + "-" + endMonth + "-" + "1";
